Check for variables that need masking/scaling.
xarray.open_zarr() supports this behavior but Xarray-TensorStore does not. So we raise an error if the dataset contains any data variables that would need masking or scaling.

We also add a mask_and_scale parameter to xarray_tensorstore.open_zarr() so that users can explicitly masking/scaling and thus avoid this error.

PiperOrigin-RevId: 646269657