[xarray-tensorstore] fix PyPI metadata and update GitHub actions
"Apache-2.0" is the standard SPDX string which PyPI prefers.
"source" is the right way to indicate a GitHub link.

We need to update the GitHub actions in order to avoid errors from running deprecated versions, but the tests from this PR are still failing on some versions of Python -- specifically those for which Zarr-Python v3 is installed. We'll fix those issues in a follow-on.

PiperOrigin-RevId: 794258008